+++ sfml-flow/src/toolgrid.h
@@ -40,7 +40,7 @@ private:
     const static int kXOffsetSize = 170;
     const static int kYOffsetSize = 40;
     const static int kButtonSize = 160;
-    const static int kMaxCharacters = 20;
+    const static size_t kMaxCharacters = 20;
     float offset_x_;
     float offset_y_;
     u_int32_t column_;
